clinical_bench/server/requirements.txt
# Server dependencies
openenv-core>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
websockets>=13.0
httpx>=0.27.0
